+++ services/dbaas/mariadb/backups.go
@@ -10,11 +10,12 @@ import (
 	mariadb "github.com/ionos-cloud/sdk-go-bundle/products/dbaas/mariadb/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/shared"
 	"github.com/ionos-cloud/sdk-go-bundle/shared/fileconfiguration"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils/loadedconfig"
 )
 
 // GetClusterBackups retrieves a list of backups for a given cluster ID and the location in which the cluster is created.
 func (c *Client) GetClusterBackups(ctx context.Context, clusterID, location string) (mariadb.BackupList, *shared.APIResponse, error) {
-	c.overrideClientEndpoint(fileconfiguration.Mariadb, location)
+	loadedconfig.SetClientOptionsFromConfig(c, fileconfiguration.Mariadb, location)
 	backups, apiResponse, err := c.sdkClient.BackupsApi.ClusterBackupsGet(ctx, clusterID).Execute()
 	apiResponse.LogInfo()
 	return backups, apiResponse, err
@@ -22,7 +23,7 @@ func (c *Client) GetClusterBackups(ctx context.Context, clusterID, location stri
 
 // FindBackupByID retrieves a backup by its ID and the location in which the cluster is created.
 func (c *Client) FindBackupByID(ctx context.Context, backupID, location string) (mariadb.BackupResponse, *shared.APIResponse, error) {
-	c.overrideClientEndpoint(fileconfiguration.Mariadb, location)
+	loadedconfig.SetClientOptionsFromConfig(c, fileconfiguration.Mariadb, location)
 	backups, apiResponse, err := c.sdkClient.BackupsApi.BackupsFindById(ctx, backupID).Execute()
 	apiResponse.LogInfo()
 	return backups, apiResponse, err