+++ services/dbaas/mariadb/cluster.go
@@ -13,11 +13,12 @@ import (
 
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils/constant"
+	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils/loadedconfig"
 )
 
 // GetCluster retrieves a cluster by its ID and the location in which the cluster is created.
 func (c *Client) GetCluster(ctx context.Context, clusterID, location string) (mariadb.ClusterResponse, *shared.APIResponse, error) {
-	c.overrideClientEndpoint(fileconfiguration.Mariadb, location)
+	loadedconfig.SetClientOptionsFromConfig(c, fileconfiguration.Mariadb, location)
 	cluster, apiResponse, err := c.sdkClient.ClustersApi.ClustersFindById(ctx, clusterID).Execute()
 	apiResponse.LogInfo()
 	return cluster, apiResponse, err
@@ -25,7 +26,7 @@ func (c *Client) GetCluster(ctx context.Context, clusterID, location string) (ma
 
 // ListClusters retrieves a list of clusters based on the location. Filters can be used.
 func (c *Client) ListClusters(ctx context.Context, filterName, location string) (mariadb.ClusterList, *shared.APIResponse, error) {
-	c.overrideClientEndpoint(fileconfiguration.Mariadb, location)
+	loadedconfig.SetClientOptionsFromConfig(c, fileconfiguration.Mariadb, location)
 	request := c.sdkClient.ClustersApi.ClustersGet(ctx)
 	if filterName != "" {
 		request = request.FilterName(filterName)
@@ -37,7 +38,7 @@ func (c *Client) ListClusters(ctx context.Context, filterName, location string) 
 
 // CreateCluster creates a new cluster using the provided data in the request and the location.
 func (c *Client) CreateCluster(ctx context.Context, cluster mariadb.CreateClusterRequest, location string) (mariadb.ClusterResponse, *shared.APIResponse, error) {
-	c.overrideClientEndpoint(fileconfiguration.Mariadb, location)
+	loadedconfig.SetClientOptionsFromConfig(c, fileconfiguration.Mariadb, location)
 	clusterResponse, apiResponse, err := c.sdkClient.ClustersApi.ClustersPost(ctx).CreateClusterRequest(cluster).Execute()
 	apiResponse.LogInfo()
 	return clusterResponse, apiResponse, err
@@ -45,7 +46,7 @@ func (c *Client) CreateCluster(ctx context.Context, cluster mariadb.CreateCluste
 
 // UpdateCluster updates a cluster by its ID and the location in which the cluster is created.
 func (c *Client) UpdateCluster(ctx context.Context, cluster mariadb.PatchClusterRequest, clusterID, location string) (mariadb.ClusterResponse, *shared.APIResponse, error) {
-	c.overrideClientEndpoint(fileconfiguration.Mariadb, location)
+	loadedconfig.SetClientOptionsFromConfig(c, fileconfiguration.Mariadb, location)
 	clusterResponse, apiResponse, err := c.sdkClient.ClustersApi.ClustersPatch(ctx, clusterID).PatchClusterRequest(cluster).Execute()
 	apiResponse.LogInfo()
 	return clusterResponse, apiResponse, err
@@ -53,7 +54,7 @@ func (c *Client) UpdateCluster(ctx context.Context, cluster mariadb.PatchCluster
 
 // DeleteCluster deletes a cluster by its ID and the location in which the cluster is created.
 func (c *Client) DeleteCluster(ctx context.Context, clusterID, location string) (mariadb.ClusterResponse, *shared.APIResponse, error) {
-	c.overrideClientEndpoint(fileconfiguration.Mariadb, location)
+	loadedconfig.SetClientOptionsFromConfig(c, fileconfiguration.Mariadb, location)
 	clusterResponse, apiResponse, err := c.sdkClient.ClustersApi.ClustersDelete(ctx, clusterID).Execute()
 	apiResponse.LogInfo()
 	return clusterResponse, apiResponse, err